% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deft-package.R
\docType{package}
\name{deft-package}
\alias{deft}
\alias{deft-package}
\title{deft: Helpful Things for Package Development}
\description{
Collection of things that I find helpful.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jamiegaida/deft}
}

}
\author{
\strong{Maintainer}: Jamie Gaida \email{jamie.gaida.codes@gmail.com} (\href{https://orcid.org/0000-0002-4504-9188}{ORCID})

}
\keyword{internal}
