% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{deft_benchmark}
\alias{deft_benchmark}
\title{A benchmarking demo which creates a data.frame (and populates it) totally 
within C}
\usage{
deft_benchmark(n = 8)
}
\arguments{
\item{n}{number of rows to generate. Default: 8}
}
\value{
a data.frame created within C
}
\description{
A benchmarking demo which creates a data.frame (and populates it) totally 
within C
}
