% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_helpers.R
\name{double_bracket_for_dot_data}
\alias{double_bracket_for_dot_data}
\title{Switch \code{.data$colname} to \code{.data[["col_name"]]}}
\usage{
double_bracket_for_dot_data(
  files = list.files(path = ".", pattern = ".R$", recursive = TRUE, full.names = TRUE)
)
}
\arguments{
\item{files}{(character vector) Full path to the files to be changed}
}
\value{
Nothing, called for side effect
}
\description{
When creating a package the preferred extraction operator is
\code{.data[["col_name"]]}. What to do if you have the form
\code{.data$col_name} littered throughout your package? Can you be bothered
finding and correcting them all?
}
\details{
This function will modify files in your package. It is strongly recommended
to save a copy of your files or use git before diving in and using this
function.
}
\examples{
\dontrun{
double_bracket_operator_for_dot_data(files = "./R/get_data.R")
}
}
