% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_functions}
\alias{load_functions}
\title{Load all function definitions without creating a project}
\usage{
load_functions(
  files = list.files(path = "R", pattern = "\\\\.R$", full.names = TRUE, recursive =
    TRUE),
  clean_up = TRUE
)
}
\arguments{
\item{files}{(character vector) Files from which to load function definitions}

\item{clean_up}{(boolean; default = TRUE) Should global function definitions
be removed before loading?}
}
\value{
Nothing, called for side effect
}
\description{
\code{devtools::load_all()} only works inside a package. \code{source()} does
the job so long as your source file only includes function definition. This
filters out anything that doesn't match the pattern \code{name <- function()}
}
\examples{
\dontrun{
load_functions()
}
}
