% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_setup.R
\name{set_rstudio_preferences}
\alias{set_rstudio_preferences}
\title{Set standard RStudio preferences}
\usage{
set_rstudio_preferences(show_invisibles = TRUE)
}
\arguments{
\item{show_invisibles}{(Boolean; default = TRUE) Ok, not everyone wants to
see invisible characters. Change it to your preference}
}
\value{
Nothing, called for side effect
}
\description{
A very opinionated set of RStudio settings
}
\examples{
\dontrun{
set_rstudio_defaults()
}
}
